casi el debug

+++ screen.h
@@ -63,7 +63,7 @@ void clear_screen_portion();
 void print_Lshift();
 void print_Rshift();
 void print_Y();
-void print_registros(uint eax, uint ebx, uint ecx, uint edx, uint esi, uint edi, uint ebp, uint esp, uint eip, short cs, short ds, short es, short fs, short gs, short ss, uint eflags);
+void print_registros(uint eax, uint ebx, uint ecx, uint edx, uint esi, uint edi, uint ebp, uint esp, uint eip, uchar cs, uchar ds, uchar es, uchar fs, uchar gs, uchar ss, uint eflags);
 void print_puntaje(jugador_t *j);
 void limpiar_clock(pirata_t *p);
 #endif  /* !__SCREEN_H__ */
